minor: compatibility for windows

+++ libraries/terry/src/terry/sampler/details.hpp
@@ -60,9 +60,9 @@ struct add_dst_mul_src
  *       ^..... loc is pointing to D point
  */
 template < typename xy_locator, typename SrcP >
-void getPixelsPointers( const xy_locator& loc, const point2<std::ptrdiff_t>& p0, const ssize_t& windowWidth, const ssize_t& imageWidth, const EParamFilterOutOfImage& outOfImageProcess, std::vector< SrcP >& src )
+void getPixelsPointers( const xy_locator& loc, const point2<std::ptrdiff_t>& p0, const int& windowWidth, const int& imageWidth, const EParamFilterOutOfImage& outOfImageProcess, std::vector< SrcP >& src )
 {
-	ssize_t middlePosition = floor( (src.size() - 1) * 0.5 );
+	int middlePosition = floor( (src.size() - 1) * 0.5 );
 
 	if( ( p0.x < 0 )  || ( p0.x > imageWidth - 1 ) )
 	{
@@ -96,7 +96,7 @@ void getPixelsPointers( const xy_locator& loc, const point2<std::ptrdiff_t>& p0,
 	}
 
 	// from center to left
-	for( ssize_t i = middlePosition - 1; i > -1; i-- )
+	for( int i = middlePosition - 1; i > -1; i-- )
 	{
 		if( ( p0.x - (middlePosition - i) > -1 ) )
 		{
@@ -160,7 +160,7 @@ void getPixelsPointers( const xy_locator& loc, const point2<std::ptrdiff_t>& p0,
 	}
 
 	// from center to right
-	for( ssize_t i = middlePosition + 1; i < (ssize_t) src.size(); i++ )
+	for( int i = middlePosition + 1; i < (int) src.size(); i++ )
 	{
 		if( ( p0.x - (middlePosition - i) < imageWidth ) )
 		{
@@ -269,7 +269,7 @@ bool process2Dresampling( Sampler& sampler, const SrcView& src, const point2<F>&
 
 	// first process the middle point
 	// if it's mirrored, we need to copy the center point
-	if( (pTL.y < 0) || (pTL.y > (ssize_t) ( src.height( ) - 1.0 ) ) )
+	if( (pTL.y < 0) || (pTL.y > (int) ( src.height( ) - 1.0 ) ) )
 	{
 		if( pTL.y < 0 ) // under the image
 		{
@@ -338,11 +338,11 @@ bool process2Dresampling( Sampler& sampler, const SrcView& src, const point2<F>&
 	}
 
 	// from center to bottom
-	for( ssize_t i = middlePosition - 1; i > -1; i-- )
+	for( int i = middlePosition - 1; i > -1; i-- )
 	{
-		if( (ssize_t) ( pTL.y - (middlePosition - i) ) < (ssize_t) src.height( ) )
+		if( (int) ( pTL.y - (middlePosition - i) ) < (int) src.height( ) )
 		{
-			if( (ssize_t) ( pTL.y - (middlePosition - i) ) < 0 )
+			if( (int) ( pTL.y - (middlePosition - i) ) < 0 )
 			{
 				switch( outOfImageProcess )
 				{
@@ -405,11 +405,11 @@ bool process2Dresampling( Sampler& sampler, const SrcView& src, const point2<F>&
 	}
 
 	// from center to top
-	for( ssize_t i = middlePosition + 1; i < (ssize_t)windowSize; i++ )
+	for( int i = middlePosition + 1; i < (int)windowSize; i++ )
 	{
-		if( (ssize_t) ( pTL.y + (i - middlePosition) ) < (ssize_t) src.height( ) )
+		if( (int) ( pTL.y + (i - middlePosition) ) < (int) src.height( ) )
 		{
-			if( (ssize_t) ( pTL.y + (i - middlePosition) )  < 0 )
+			if( (int) ( pTL.y + (i - middlePosition) )  < 0 )
 			{
 				xProcessed.at( i ) = xProcessed.at( i - 1 );
 			}
